Add ability to cache docker images
